Make control the timeout until a WebRTC connection is established
Set the timeout from the start of signaling until it is connected with
OvenMediaEngine. connectionTimeout sets the maximum allowable time
until a connection is established. timeoutMaxRetry sets the number of
times the player will automatically retry the connection when the
maximum allowed time has elapsed. When retrying a connection due to a
timeout, the player does not issue a separate error message. If the
connection fails after retries for timeoutMaxRetry, the player throws a
timeout error. If timeoutMaxRetry is set to 0, no timeout processing is
performed.